+++ packages/webawesome/src/components/details/details.ts
@@ -114,7 +114,7 @@ export default class WaDetails extends WebAwesomeElement {
   private handleSummaryClick(event: MouseEvent) {
     let targetElement = getTargetElement(event);
 
-    if (targetElement?.closest('a, button, wa-button, input, wa-input, textarea, wa-textarea, select, wa-select')) {
+    if (targetElement?.closest('a, button, wa-button, input, wa-input, textarea, wa-textarea')) {
       // Let interactive elements handle their own clicks, fixes #309
       return;
     }